+++ changelog.md
@@ -2,7 +2,10 @@
 
 ##### Breaking Changes
 
-* None.
+* Updated the WorkPackageConfig to include ResultsConfig
+* The ModelConfig in a WorkPackageConfig is not optional anymore
+* The ModelConfig now includes load time information
+* The ModelConfig now includes a flag to signal the model is a calibration model
 
 ##### New Features